fix(Migration): also migrate permanent lists
Permanent lists are used also as base for variable

+++ inkcpp/array.h
@@ -20,7 +20,7 @@ namespace ink::runtime::internal
  * @tparam simple if the object has a trivial destructor, so delete[](char*) can be used instead of
  * calling the constructor.
  * @tparam dynamic if the memory should be allocated on the heap and grow if needed
- * @tparam initialCapacitiy number of elements to allocate at construction, if !dynamic, this is
+ * @tparam initial capacity number of elements to allocate at construction, if !dynamic, this is
  * allocated in place and can not be changed.
  */
 template<typename T, bool dynamic, size_t initialCapacity, bool simple>
@@ -312,7 +312,7 @@ public:
 		clear_temp();
 	}
 
-	// == Non-Copyable ==
+	// not copyable
 	basic_restorable_array(const basic_restorable_array<T>&)               = delete;
 	basic_restorable_array<T>& operator=(const basic_restorable_array<T>&) = delete;
 
@@ -372,15 +372,15 @@ private:
 	// real values live here
 	T* _array;
 
-	// we store values here when we're in save mode
+	// we store values here when we're in safe mode
 	//  they're copied on a call to forget()
 	T* _temp;
 
 	// size of both _array and _temp
 	size_t _capacity;
-	// if loaded with snap_load, this value was the original size, the current capacaty might be
+	// if loaded with snap_load, this value was the original size, the current capacity might be
 	// higher
-	size_t _loaded_capacity = ~0;
+	size_t _loaded_capacity = static_cast<size_t>(~0);
 
 	// null
 	const T _null;
@@ -406,7 +406,7 @@ inline const T& basic_restorable_array<T>::get(size_t index) const
 {
 	check_index(index);
 
-	// If we're in save mode and we have a value at that index, return that instead
+	// If we're in safe mode, and we have a value at that index, return that instead
 	if (_saved && _temp[index] != _null) {
 		return _temp[index];
 	}
@@ -445,7 +445,7 @@ inline void basic_restorable_array<T>::forget()
 {
 	// Run through the _temp array
 	for (size_t i = 0; i < _capacity; i++) {
-		// Copy if there's values
+		// Copy if there are values
 		if (_temp[i] != _null) {
 			_array[i] = _temp[i];
 		}
